refactor(effects): classify depth of field as screen-space
