--- modules/effects/src/passes/screen-space/types.ts
@@ -1,5 +1,0 @@
-// luma.gl
-// SPDX-License-Identifier: MIT
-// Copyright (c) vis.gl contributors
-
-export type ScreenSpaceNormalSource = 'reconstruct-from-depth' | 'normal-texture';
