+++ modules/effects/test/index.ts
@@ -5,7 +5,7 @@
 // Post-processing Modules (Shader passes)
 // Blur Filters
 import './passes/image-blur-filters/bloom.spec';
-import './passes/image-blur-filters/dof.spec';
+import './passes/screen-space/dof.spec';
 import './passes/image-blur-filters/gaussianblur.spec';
 import './passes/image-blur-filters/tiltshift.spec';
 import './passes/image-blur-filters/triangleblur.spec';
